Give the hash source a module of its own
`misc` is where a thing with no home goes. The hash source has one: the hashes
the compiler names things by.
